More region renaming
RecordSensorFilter -> record_sensor_filters
SVMClassifierNode -> svm_classifier_node
TestRegion -> test_region
UnimportableNode -> unimportable_node
