+++ examples/prediction/experiments/dutyCycle/base/description.py
@@ -35,7 +35,7 @@ from nupic.frameworks.prediction.callbacks import (printSPCoincidences,
                                                    displaySPCoincidences,
                                                    setAttribute,
                                                    sensorOpen)
-from nupic.regions.RecordSensorFilters.modify_fields import ModifyFields
+from nupic.regions.record_sensor_filters.modify_fields import ModifyFields
