+++ src/nupic/engine/__init__.py
@@ -52,10 +52,10 @@ pyRegions = (
     ("nupic.regions.record_sensor", "RecordSensor"),
     ("nupic.regions.sdr_classifier_region", "SDRClassifierRegion"),
     ("nupic.regions.sp_region", "SPRegion"),
-    ("nupic.regions.SVMClassifierNode", "SVMClassifierNode"),
-    ("nupic.regions.TestRegion", "TestRegion"),
+    ("nupic.regions.svm_classifier_node", "SVMClassifierNode"),
+    ("nupic.regions.test_region", "TestRegion"),
     ("nupic.regions.tm_region", "TMRegion"),
-    ("nupic.regions.UnimportableNode", "UnimportableNode"),
+    ("nupic.regions.unimportable_node", "UnimportableNode"),
 )
 
 registeredRegions = False
