+++ src/nupic/regions/record_sensor_filters/add_noise.py
