+++ src/nupic/regions/record_sensor_filters/modify_fields.py
