+++ src/nupic/regions/svm_classifier_node.py
